web-backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==2.0.40
psycopg2-binary==2.9.9
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
python-dotenv==1.0.1
requests
httpx[http2]==0.27.0